+++ src/test/java/org/ojalgo/optimisation/integer/MipCase.java
@@ -24,7 +24,8 @@ package org.ojalgo.optimisation.integer;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
-import org.ojalgo.optimisation.ModelFileMPS;
+import org.ojalgo.optimisation.ExpressionsBasedModel.FileFormat;
+import org.ojalgo.optimisation.ModelFileTest;
 import org.ojalgo.type.context.NumberContext;
 
 /**
@@ -35,12 +36,12 @@ import org.ojalgo.type.context.NumberContext;
  */
 @Disabled("Too slow")
 @Tag("slow")
-public final class MipCase extends OptimisationIntegerTests implements ModelFileMPS {
+public final class MipCase extends OptimisationIntegerTests implements ModelFileTest {
 
-    static final NumberContext PRECISION = NumberContext.of(8, 6);
+    private static final NumberContext ACCURACY = NumberContext.of(8, 6);
 
-    static void doTest(final String modelName, final String expMinValString, final String expMaxValString) {
-        ModelFileMPS.makeAndAssert("miplib", modelName, expMinValString, expMaxValString, false, PRECISION, null);
+    private static void doTest(final String modelName, final String expMinValString, final String expMaxValString) {
+        ModelFileTest.makeAndAssert("miplib", modelName, FileFormat.MPS, false, expMinValString, expMaxValString, ACCURACY);
     }
 
     /**